Celebrate 20th century female Mathematicians and Computer Scientists (31 new)
The following were already on the list from 19th century women:
* Hertha Ayrton

The following did not have a birth date listed:
* Mandy Chessel
* Shafi Goldwasser
* Bryna Kra
* Julia Hirschberg
* Marta Kwiatkowska
* Mangala Narlikar
* Jeannette Wing

Special prop on Betty Holberton and Grace Hopper about "inventions"
(breakpoints in programming, and COBOL). Considering a reason to
celebrate someone is what they've invented or brought to the world, as
many inventors are less known than the inventions.

https://en.wikipedia.org/wiki/List_of_female_scientists_in_the_20th_century#Mathematics_or_computer_science